Removing SingletonDriver and reorganising new util packages
